Move icons into a folder
[MAILPOET-2742]
