+++ assets/js/src/form_editor/components/form_settings/form_placement_options/fixed_bar.tsx
@@ -5,7 +5,7 @@ import { useSelect, useDispatch } from '@wordpress/data';
 
 import FormPlacementSettings from './form_placement_settings';
 import Toggle from '../../../../common/toggle';
-import Icon from './fixed_bar_icon';
+import Icon from './icons/fixed_bar_icon';
 
 const delayValues = [0, 15, 30, 60, 120, 180, 240];
