+++ assets/js/src/form_editor/components/form_settings/form_placement_options/icons/fixed_bar_icon.tsx
