+++ assets/js/src/form_editor/components/form_settings/form_placement_options/icons/popup_icon.tsx
