+++ assets/js/src/form_editor/components/form_settings/form_placement_options/icons/settings_icon.tsx
