+++ assets/js/src/form_editor/components/form_settings/form_placement_options/popup.tsx
@@ -5,7 +5,7 @@ import { useSelect, useDispatch } from '@wordpress/data';
 
 import FormPlacementSettings from './form_placement_settings';
 import Toggle from '../../../../common/toggle';
-import Icon from './popup_icon';
+import Icon from './icons/popup_icon';
 
 const delayValues = [0, 15, 30, 60, 120, 180, 240];
